requirements.txt
Django==1.10.2
django-tastypie==0.13.3
python-dateutil==2.5.3
python-mimeparse==1.6.0
six==1.10.0
